+++ src/components/nav.js
@@ -1,6 +1,9 @@
-import React from "react";
+import React, { useContext } from "react";
+import { MenuContext } from "react-flexible-sliding-menu";
 
 const Nav = () => {
+  const { toggleMenu } = useContext(MenuContext);
+
   return (
     <div className="nav fixed">
       <nav id="nav">
@@ -12,21 +15,9 @@ const Nav = () => {
           <h2>Krity Dangol</h2>
           <p>Front-end Developer</p>
         </a>
-
-        <label>
-          <input type="checkbox" />
-          <span className="menu">
-            <span className="hamburger"></span>
-          </span>
-          <ul id="nav-links">
-            <li>
-              <a href="/">Home</a>
-            </li>
-            <li>
-              <a href="/home">Home real</a>
-            </li>
-          </ul>
-        </label>
+        <button onClick={toggleMenu} className="primary-button">
+          Toggle Menu
+        </button>
       </nav>
     </div>
   );